requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn==0.30.6
